Move service classes to top-level package.
