+++ service/src/main/java/bio/terra/tanagra/app/controller/EntitiesV2ApiController.java
@@ -7,9 +7,9 @@ import bio.terra.tanagra.generated.controller.EntitiesV2Api;
 import bio.terra.tanagra.generated.model.ApiEntityListV2;
 import bio.terra.tanagra.generated.model.ApiEntityV2;
 import bio.terra.tanagra.service.AccessControlService;
-import bio.terra.tanagra.service.ToApiConversionUtils;
 import bio.terra.tanagra.service.UnderlaysService;
 import bio.terra.tanagra.service.accesscontrol.ResourceId;
+import bio.terra.tanagra.service.utils.ToApiConversionUtils;
 import bio.terra.tanagra.underlay.Entity;
 import java.util.stream.Collectors;
 import org.springframework.beans.factory.annotation.Autowired;
