+++ service/src/main/java/bio/terra/tanagra/app/controller/HintsV2ApiController.java
@@ -16,10 +16,10 @@ import bio.terra.tanagra.query.Literal;
 import bio.terra.tanagra.query.QueryRequest;
 import bio.terra.tanagra.service.AccessControlService;
 import bio.terra.tanagra.service.FromApiConversionService;
-import bio.terra.tanagra.service.ToApiConversionUtils;
+import bio.terra.tanagra.service.QuerysService;
 import bio.terra.tanagra.service.UnderlaysService;
 import bio.terra.tanagra.service.accesscontrol.ResourceId;
-import bio.terra.tanagra.service.instances.QuerysService;
+import bio.terra.tanagra.service.utils.ToApiConversionUtils;
 import bio.terra.tanagra.underlay.Attribute;
 import bio.terra.tanagra.underlay.DisplayHint;
 import bio.terra.tanagra.underlay.Entity;
