+++ service/src/main/java/bio/terra/tanagra/app/controller/InstancesV2ApiController.java
@@ -19,15 +19,15 @@ import bio.terra.tanagra.query.OrderByDirection;
 import bio.terra.tanagra.query.QueryRequest;
 import bio.terra.tanagra.service.AccessControlService;
 import bio.terra.tanagra.service.FromApiConversionService;
-import bio.terra.tanagra.service.ToApiConversionUtils;
+import bio.terra.tanagra.service.QuerysService;
 import bio.terra.tanagra.service.UnderlaysService;
 import bio.terra.tanagra.service.accesscontrol.ResourceId;
-import bio.terra.tanagra.service.filter.EntityFilter;
 import bio.terra.tanagra.service.instances.EntityInstance;
 import bio.terra.tanagra.service.instances.EntityInstanceCount;
 import bio.terra.tanagra.service.instances.EntityQueryOrderBy;
 import bio.terra.tanagra.service.instances.EntityQueryRequest;
-import bio.terra.tanagra.service.instances.QuerysService;
+import bio.terra.tanagra.service.instances.filter.EntityFilter;
+import bio.terra.tanagra.service.utils.ToApiConversionUtils;
 import bio.terra.tanagra.underlay.Attribute;
 import bio.terra.tanagra.underlay.DataPointer;
 import bio.terra.tanagra.underlay.Entity;
