+++ service/src/main/java/bio/terra/tanagra/app/controller/StudiesV2ApiController.java
@@ -14,10 +14,10 @@ import bio.terra.tanagra.generated.model.ApiStudyListV2;
 import bio.terra.tanagra.generated.model.ApiStudyUpdateInfoV2;
 import bio.terra.tanagra.generated.model.ApiStudyV2;
 import bio.terra.tanagra.service.AccessControlService;
+import bio.terra.tanagra.service.StudyService;
 import bio.terra.tanagra.service.accesscontrol.ResourceId;
 import bio.terra.tanagra.service.accesscontrol.ResourceIdCollection;
 import bio.terra.tanagra.service.artifact.Study;
-import bio.terra.tanagra.service.artifact.StudyService;
 import com.google.common.collect.ImmutableMap;
 import java.util.HashMap;
 import java.util.List;
