+++ service/src/main/java/bio/terra/tanagra/service/FromApiConversionService.java
@@ -13,17 +13,16 @@ import bio.terra.tanagra.query.Literal;
 import bio.terra.tanagra.query.filtervariable.BinaryFilterVariable;
 import bio.terra.tanagra.query.filtervariable.BooleanAndOrFilterVariable;
 import bio.terra.tanagra.query.filtervariable.FunctionFilterVariable;
-import bio.terra.tanagra.service.filter.AttributeFilter;
-import bio.terra.tanagra.service.filter.BooleanAndOrFilter;
-import bio.terra.tanagra.service.filter.BooleanNotFilter;
-import bio.terra.tanagra.service.filter.EntityFilter;
-import bio.terra.tanagra.service.filter.HierarchyAncestorFilter;
-import bio.terra.tanagra.service.filter.HierarchyMemberFilter;
-import bio.terra.tanagra.service.filter.HierarchyParentFilter;
-import bio.terra.tanagra.service.filter.HierarchyRootFilter;
-import bio.terra.tanagra.service.filter.RelationshipFilter;
-import bio.terra.tanagra.service.filter.TextFilter;
-import bio.terra.tanagra.service.instances.QuerysService;
+import bio.terra.tanagra.service.instances.filter.AttributeFilter;
+import bio.terra.tanagra.service.instances.filter.BooleanAndOrFilter;
+import bio.terra.tanagra.service.instances.filter.BooleanNotFilter;
+import bio.terra.tanagra.service.instances.filter.EntityFilter;
+import bio.terra.tanagra.service.instances.filter.HierarchyAncestorFilter;
+import bio.terra.tanagra.service.instances.filter.HierarchyMemberFilter;
+import bio.terra.tanagra.service.instances.filter.HierarchyParentFilter;
+import bio.terra.tanagra.service.instances.filter.HierarchyRootFilter;
+import bio.terra.tanagra.service.instances.filter.RelationshipFilter;
+import bio.terra.tanagra.service.instances.filter.TextFilter;
 import bio.terra.tanagra.underlay.Entity;
 import bio.terra.tanagra.underlay.EntityGroup;
 import bio.terra.tanagra.underlay.Hierarchy;
