+++ service/src/main/java/bio/terra/tanagra/service/instances/filter/HierarchyRootFilter.java
@@ -1,4 +1,4 @@
-package bio.terra.tanagra.service.filter;
+package bio.terra.tanagra.service.instances.filter;
 
 import bio.terra.tanagra.query.FieldVariable;
 import bio.terra.tanagra.query.FilterVariable;
@@ -10,10 +10,10 @@ import bio.terra.tanagra.underlay.HierarchyField;
 import bio.terra.tanagra.underlay.Underlay;
 import java.util.List;
 
-public class HierarchyMemberFilter extends EntityFilter {
+public class HierarchyRootFilter extends EntityFilter {
   private final Hierarchy hierarchy;
 
-  public HierarchyMemberFilter(Hierarchy hierarchy) {
+  public HierarchyRootFilter(Hierarchy hierarchy) {
     this.hierarchy = hierarchy;
   }
 
@@ -25,8 +25,8 @@ public class HierarchyMemberFilter extends EntityFilter {
         pathField.buildFieldVariableFromEntityId(
             hierarchy.getMapping(Underlay.MappingType.INDEX), entityTableVar, tableVars);
 
-    // IS_MEMBER translates to path IS NOT NULL
+    // IS_ROOT translates to path=""
     return new BinaryFilterVariable(
-        pathFieldVar, BinaryFilterVariable.BinaryOperator.IS_NOT, new Literal(null));
+        pathFieldVar, BinaryFilterVariable.BinaryOperator.EQUALS, new Literal(""));
   }
 }
