+++ service/src/main/java/bio/terra/tanagra/service/utils/ToApiConversionUtils.java
@@ -1,4 +1,4 @@
-package bio.terra.tanagra.service;
+package bio.terra.tanagra.service.utils;
 
 import bio.terra.tanagra.exception.SystemException;
 import bio.terra.tanagra.generated.model.ApiAttributeV2;
