+++ service/src/test/java/bio/terra/tanagra/api/BaseQueriesTest.java
@@ -6,18 +6,18 @@ import bio.terra.tanagra.query.Literal;
 import bio.terra.tanagra.query.QueryRequest;
 import bio.terra.tanagra.query.filtervariable.BooleanAndOrFilterVariable;
 import bio.terra.tanagra.query.filtervariable.FunctionFilterVariable;
+import bio.terra.tanagra.service.QuerysService;
 import bio.terra.tanagra.service.UnderlaysService;
-import bio.terra.tanagra.service.filter.AttributeFilter;
-import bio.terra.tanagra.service.filter.BooleanAndOrFilter;
-import bio.terra.tanagra.service.filter.EntityFilter;
-import bio.terra.tanagra.service.filter.HierarchyAncestorFilter;
-import bio.terra.tanagra.service.filter.HierarchyMemberFilter;
-import bio.terra.tanagra.service.filter.HierarchyParentFilter;
-import bio.terra.tanagra.service.filter.HierarchyRootFilter;
-import bio.terra.tanagra.service.filter.RelationshipFilter;
-import bio.terra.tanagra.service.filter.TextFilter;
 import bio.terra.tanagra.service.instances.EntityQueryRequest;
-import bio.terra.tanagra.service.instances.QuerysService;
+import bio.terra.tanagra.service.instances.filter.AttributeFilter;
+import bio.terra.tanagra.service.instances.filter.BooleanAndOrFilter;
+import bio.terra.tanagra.service.instances.filter.EntityFilter;
+import bio.terra.tanagra.service.instances.filter.HierarchyAncestorFilter;
+import bio.terra.tanagra.service.instances.filter.HierarchyMemberFilter;
+import bio.terra.tanagra.service.instances.filter.HierarchyParentFilter;
+import bio.terra.tanagra.service.instances.filter.HierarchyRootFilter;
+import bio.terra.tanagra.service.instances.filter.RelationshipFilter;
+import bio.terra.tanagra.service.instances.filter.TextFilter;
 import bio.terra.tanagra.testing.BaseSpringUnitTest;
 import bio.terra.tanagra.testing.GeneratedSqlUtils;
 import bio.terra.tanagra.underlay.Attribute;
