+++ service/src/test/java/bio/terra/tanagra/api/omop/OmopIngredientTest.java
@@ -3,9 +3,9 @@ package bio.terra.tanagra.api.omop;
 import bio.terra.tanagra.api.BaseQueriesTest;
 import bio.terra.tanagra.query.Literal;
 import bio.terra.tanagra.query.filtervariable.BinaryFilterVariable;
-import bio.terra.tanagra.service.filter.AttributeFilter;
-import bio.terra.tanagra.service.filter.RelationshipFilter;
 import bio.terra.tanagra.service.instances.EntityQueryRequest;
+import bio.terra.tanagra.service.instances.filter.AttributeFilter;
+import bio.terra.tanagra.service.instances.filter.RelationshipFilter;
 import bio.terra.tanagra.testing.GeneratedSqlUtils;
 import bio.terra.tanagra.underlay.Entity;
 import bio.terra.tanagra.underlay.Relationship;
