+++ service/src/test/java/bio/terra/tanagra/api/omop/OmopMeasurementTest.java
@@ -4,10 +4,10 @@ import bio.terra.tanagra.api.BaseQueriesTest;
 import bio.terra.tanagra.query.Literal;
 import bio.terra.tanagra.query.filtervariable.BinaryFilterVariable;
 import bio.terra.tanagra.query.filtervariable.BooleanAndOrFilterVariable;
-import bio.terra.tanagra.service.filter.AttributeFilter;
-import bio.terra.tanagra.service.filter.BooleanAndOrFilter;
-import bio.terra.tanagra.service.filter.HierarchyRootFilter;
 import bio.terra.tanagra.service.instances.EntityQueryRequest;
+import bio.terra.tanagra.service.instances.filter.AttributeFilter;
+import bio.terra.tanagra.service.instances.filter.BooleanAndOrFilter;
+import bio.terra.tanagra.service.instances.filter.HierarchyRootFilter;
 import bio.terra.tanagra.testing.GeneratedSqlUtils;
 import bio.terra.tanagra.underlay.Underlay;
 import java.io.IOException;
